Move modelAndMakeCarDetails component into its own folder for better modularisation
